add new automatic notify email

+++ mart/frontweb/route.py
@@ -1,16 +1,15 @@
 from datetime import timedelta
 import datetime
-import sys
-import time
-
 from celery import Celery
 from flask import *
-
-from mart import db, app
+from mart import db, app, mail
+from flask_mail import Message
 from mart.constant.appConstant import Constant
 from mart.frontweb.forms import UserPost
 from mart.models import Categories, Subcategories, Products
 from mart.models import PostUser
+from celery.schedules import crontab
+from celery_once import QueueOnce
 
 celery = Celery('hello', broker="redis://guest@localhost//", backend='redis')
 
@@ -38,19 +37,17 @@ def show_pro_id(pro_id):
 
 @front_web.route('/post_front', methods=[Constant.GET, Constant.POST])
 def front_post():
-    # print(countdown(60))
-    one_day_interval_before = datetime.datetime.now() - timedelta(minutes=1)
-    print(one_day_interval_before)
-    now = datetime.datetime.now()
-    new_now = datetime.datetime.strftime(now, '%a, %d %b %Y %H:%M:%S %Z')
-    # print(new_now)
     form = UserPost()
     user_post = PostUser.query.all()
     if form.validate_on_submit():
         post_user = PostUser(title=form.title.data, content=form.content.data)
         db.session.add(post_user)
         db.session.commit()
-        return redirect(url_for('main.home'))
+        msg = Message('Hello', sender='salmansaleem036@gmail.com', recipients=['idreesrehan234@gmail.com'])
+        msg.body = "This is the email body"
+        mail.send(msg)
+
+        return redirect(url_for('front_web.front_home'))
     return render_template('add_front_post.html', form=form, user_post=user_post)
 
 
@@ -74,50 +71,46 @@ def active_post_user(id):
 
 
 # here celery tasks
-
-
-def countdown(totalTime):
-    try:
-        while totalTime >= 0:
-            mins, secs = divmod(totalTime, 60)
-            sys.stdout.write("\rWaiting for {:02d}:{:02d}  minutes...".format(mins, secs))
-            sys.stdout.flush()
-            time.sleep(1)
-            totalTime -= 1
-            if totalTime <= -1:
-                print
-                "\n"
-                break
-    except KeyboardInterrupt:
-        exit("\n^C Detected!\nExiting...")
+@celery.task
+def send_async_email(msg):
+    """Background task to send an email with Flask-Mail."""
+    with app.app_context():
+        mail.send(msg)
 
 
 @celery.task
 def index():
     try:
-        now = datetime.datetime.now()
-        before = timedelta(minutes=1)
-        now = now.replace(microsecond=0)
-        before = (now + before)
-        now = (now.strftime("%b %d %X"))
-        # before = (before.strftime("%b %d %X"))
-        # print('Before ' + before)
-        # print('Now ' +now)
-        since = datetime.datetime.now()-timedelta(minutes=5)
-        print(since)
         post = PostUser.query.filter_by(status=False).first()
-
-        db.session.delete(post)
-        db.session.commit()
+        print(post.post_at < datetime.datetime.now() - timedelta(minutes=3))
+        if post.post_at < datetime.datetime.now() - timedelta(minutes=3):
+            print('Come here')
+            db.session.delete(post)
+            db.session.commit()
     except:
         print('empty table')
     return 'sent'
 
 
-@celery.task
-def create_db():
-    db.create_all()
-    return 'create'
+@celery.task()
+def send_email_notification():
+    try:
+        post = PostUser.query.filter_by(status=False).first()
+        print(post.post_at < datetime.datetime.now() - timedelta(minutes=1))
+        if post.post_at < datetime.datetime.now() - timedelta(minutes=1):
+            print('email send')
+            msg = Message('Hello', sender='salmansaleem036@gmail.com', recipients=['idreesrehan234@gmail.com'])
+            msg.body = "After two minutes post will expire"
+            send_async_email(msg)
+    except:
+        print('No email')
+    return 'Email send'
+
+
+@celery.task(bind=True)
+def check():
+    print('check')
+    return ''
 
 
 @celery.on_after_configure.connect
@@ -125,4 +118,15 @@ def setup_periodic_tasks(sender, **kwargs):
     sender.add_periodic_task(
         timedelta(seconds=30),
         index,
+
     )
+    sender.add_periodic_task(
+        timedelta(seconds=30),
+        send_email_notification,
+
+    )
+    # sender.add_periodic_task(
+    #     timedelta(seconds=30),
+    #
+    #     check,
+    # )
